Update patch set 1
Patch Set 1:

A similar patch has already been merged into linaro which is why this patch has such a old authorship date.  This warning with clang is a know issue, just Google search the error.

https://www.google.com/webhp?sourceid=chrome-instant&ion=1&ie=UTF-8#hl=en&sclient=psy-ab&q=clang%3A%20error%3A%20argument%20unused%20during%20compilation%20Qunused-arguments&oq=&gs_l=&pbx=1&fp=6edddecf24550d4a&ion=1&bav=on.2,or.r_cp.r_qf.&bvm=bv.45368065,d.aWc&biw=1285&bih=551

But with this patch I improved it a little more to not use as many flags.  The solution is to use -Qunused-arguments
There is no other way to fix this..

Patch-set: 1